requirements.txt
requests==2.26.0
beautifulsoup4==4.9.3
mysql_connector_repackaged==0.3.1
